requirements.txt
psycopg2-binary
graphviz
pandas
numpy
altair
google-cloud-bigquery==2.28.1
